fix: address code review — invariant DateTime parsing, internal TurnNumber setter, int RngSeed

+++ src/Scripts/Core/RunState.cs
@@ -7,7 +7,7 @@ public class RunState
     public int PlayerHealth { get; set; }
     public int EnemyHealth { get; set; }
     public int TurnCount { get; set; }
-    public long RngSeed { get; set; }
+    public int RngSeed { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime UpdatedAt { get; set; }
 }
